docker-fastapi-celery/requirements_celery.txt
celery==5.5.1
redis==5.0.1
llama-index==0.12.31
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.31
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-instructor==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-huggingface==0.5.0
llama-index-llms-huggingface-api==0.4.1
llama-index-llms-openai==0.3.37
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-postprocessor-colpali-rerank==0.2.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-storage-docstore-mongodb==0.3.0
llama-index-storage-kvstore-mongodb==0.3.0
llama-index-vector-stores-weaviate==1.3.1
slowapi==0.1.9
starlette==0.46.2
